fix: move ToggleIconButton to UI folder
