+++ packages/frontend/src/pages/campaigns/all-campaigns-section/campaigns-top-nav/index.tsx
@@ -1,6 +1,6 @@
 import { Select, SelectOption } from "@carrot-kpi/ui";
 import React, { useState } from "react";
-import { ToggleIconButton } from "../../campaigns-filters/toggle-icon-button";
+import { ToggleIconButton } from "../../../../components/ui/toggle-icon-button";
 import { ToggleFiltersButton } from "./toggle-filters-button";
 
 interface CampaignsTopNavProps {
